Don't read in bases as floats.
Fractional bases, e.g., 1/2 do not really have any meaning, due to the
properties of exponentiations. Floating point bases are even worse.

In base 1/2, 10.1 is 1 * 1/2 + 1/(1/2) = 1/2 + 2 = 2.5 (note how the value
on the right of the point is where the 2 comes from. In base 2, 2.5 is
also 10.1. But 1.1 is more problematic, since it is 1 * (1/2)^0 + 1/(1/2) =
3, which is not so helpful.

Signed-off-by: mulhern <mulhern@cs.wisc.edu>